+++ tests/data/google/returning_none/cases.py
@@ -1,0 +1,8 @@
+def func(arg1: int) -> None:
+    """
+    Do something
+
+    Args:
+        arg1 (int): Arg 1
+    """
+    print(1)