+++ tests/data/numpy/returning_none/cases.py
@@ -1,0 +1,10 @@
+def func(arg1: int) -> None:
+    """
+    Do something
+
+    Parameters
+    ----------
+    arg1 : int
+        Arg 1
+    """
+    print(1)